requirements.txt

pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras